fix: firebase login fails on Safari 16.0+ version
This patch tries to fix firebase login problem.
We do not know if this patch change fixes the problem
or not. We have to merge this commit into dev environment
and test it afterwards.

The related issue is
https://github.com/firebase/firebase-js-sdk/issues/6716

In short, the root cause is cross-origin storage access.
Therefore, this patch updates firebase config `authDomain`.